+++ tests/photons_device_finder_tests/test_device.py
@@ -248,9 +248,9 @@ describe TestCase, "Device":
             assert location_uuid2 in self.collections.collections["location"]
 
         it "takes in StateHostFirmware":
-            pkt = DeviceMessages.StateHostFirmware.empty_normalise(version=1.22)
+            pkt = DeviceMessages.StateHostFirmware.empty_normalise(version="1.20")
             self.assertIs(self.device.set_from_pkt(pkt, self.collections), InfoPoints.FIRMWARE)
-            self.assertEqual(self.device.firmware_version, 1.22)
+            self.assertEqual(self.device.firmware_version, "1.20")
 
         it "takes in StateVersion":
             pkt = DeviceMessages.StateVersion.empty_normalise(vendor=1, product=22)